move pins to different package
