+++ .gitignore
@@ -1,4 +1,3 @@
 *.local
-pinctrl
 maria@raspberrypi
 bin/
