+++ pi5/board.go
@@ -9,6 +9,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/viam-modules/pinctrl/pinctrl"
+
 	mmap "github.com/edsrzf/mmap-go"
 	"github.com/pkg/errors"
 	"go.uber.org/multierr"
@@ -110,8 +112,8 @@ func newBoard(
 		cancelCtx:    cancelCtx,
 		cancelFunc:   cancelFunc,
 
-		gpios:      map[string]*gpioPin{},
-		interrupts: map[string]*digitalInterrupt{},
+		gpios:      map[string]*pinctrl.GPIOPin{},
+		interrupts: map[string]*pinctrl.DigitalInterrupt{},
 
 		chipSize: 0x30000,
 		pulls:    map[int]byte{},
@@ -125,7 +127,7 @@ func newBoard(
 
 	// Initialize the GPIO pins
 	for newName, mapping := range gpioMappings {
-		b.gpios[newName] = b.createGpioPin(mapping)
+		b.gpios[newName] = pinctrl.CreateGpioPin(cancelCtx, mapping, &b.activeBackgroundWorkers, logger, &b.vPage)
 	}
 
 	if err := b.Reconfigure(ctx, nil, conf); err != nil {
@@ -190,20 +192,6 @@ func (b *pinctrlpi5) setPulls() {
 	}
 }
 
-func (b *pinctrlpi5) createGpioPin(mapping gl.GPIOBoardMapping) *gpioPin {
-	pin := gpioPin{
-		boardWorkers: &b.activeBackgroundWorkers,
-		devicePath:   mapping.GPIOChipDev,
-		offset:       uint32(mapping.GPIO),
-		cancelCtx:    b.cancelCtx,
-		logger:       b.logger,
-	}
-	if mapping.HWPWMSupported {
-		pin.hwPwm = newPwmDevice(mapping.PWMSysFsDir, mapping.PWMID, b.logger, &b.vPage)
-	}
-	return &pin
-}
-
 type pinctrlpi5 struct {
 	resource.Named
 	mu sync.Mutex
@@ -211,8 +199,8 @@ type pinctrlpi5 struct {
 	gpioMappings map[string]gl.GPIOBoardMapping
 	logger       logging.Logger
 
-	gpios      map[string]*gpioPin
-	interrupts map[string]*digitalInterrupt
+	gpios      map[string]*pinctrl.GPIOPin
+	interrupts map[string]*pinctrl.DigitalInterrupt
 
 	virtAddr *byte     // base address of mapped virtual page referencing the gpio chip data
 	physAddr uint64    // base address of the gpio chip data in /dev/mem/
@@ -260,7 +248,7 @@ func (b *pinctrlpi5) DigitalInterruptByName(name string) (board.DigitalInterrupt
 		Name: name,
 		Pin:  name,
 	}
-	interrupt, err := newDigitalInterrupt(defaultInterruptConfig, mapping, nil)
+	interrupt, err := pinctrl.NewDigitalInterrupt(defaultInterruptConfig, mapping, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -317,9 +305,9 @@ func (b *pinctrlpi5) SetPowerMode(
 func (b *pinctrlpi5) StreamTicks(ctx context.Context, interrupts []board.DigitalInterrupt, ch chan board.Tick,
 	extra map[string]interface{},
 ) error {
-	var rawInterrupts []*digitalInterrupt
+	var rawInterrupts []*pinctrl.DigitalInterrupt
 	for _, i := range interrupts {
-		raw, ok := i.(*digitalInterrupt)
+		raw, ok := i.(*pinctrl.DigitalInterrupt)
 		if !ok {
 			return errors.New("cannot stream ticks to an interrupt not associated with this board")
 		}
